regression-tests.dnsdist/requirements.txt
dnspython>=1.11,<1.16.0
nose>=1.3.7
libnacl>=1.4.3
requests>=2.1.0
protobuf>=3.0
pysnmp>=4.3.4
future>=0.17.1
pycurl>=7.43.0
